requirements.txt
requests
azure-functions
pandas
numpy
pymssql
